Implement proof of concept GrpcWriteChannel
This is a basic implementation of inserts using gRPC. There are a bunch
of TODOs, but I've tested it against the gRPC fake server, and it
writes successfully for a range of file sizes.

The proto files were copied from google3 but with the annotations
stripped out. I'm not sure if there's a better way to do that, or how
we'll ultimately keep those in sync, but I'm assuming there's a standard
approach we can use.

I'm not sure to what degree `AbstractGoogleAsyncWriteChannel` is a
public interface that needs to stay as is. I've done my best to change
as little as possible there, and I've put most of its behavior into
BaseAbstractGoogleAsyncWriteChannel with only the JSON API specific
code staying in AbstractGoogleAsyncWriteChannel. I'm not 100% happy
with the inheritance structure - any suggestions are welcome!

Change-Id: I85cf5965606cf2123854221ad9ef4afc123b6bbf