Fix bug in GrpcWriteChannel
I found this one when I wrote the tests for the write path.

The writeChannel was always sending chunks of 8MB (even when there was
less data). Since the remaining bytes were null characters they were
getting stripped out by the server, so the result was correct, but
that's a pretty silly waste of bandwidth!

Change-Id: I9eedd65819ef7c738d3a06b2098b4de48695c14b