Validate object crc32c checksums on gRPC read
This commit only validates the object checksums. Per-message checksum
support will follow once the server supoorts them.

We can only validate object checksums in the case of a complete
read of the object. This commit only aims to validate object checksums
for complete *sequential* reads, potentially over multiple calls to
`read`, but with a single call to `open`.

This should allow object checksums to be validated for common cases
where the user reads the whole file such as a hadoop
`FileSystem.copyToLocalFile` call.

We do not attempt to validate object checksums if Fadvise.RANDOM or
GenerationReadConsistency.LATEST is set. While in principle we could
sometimes detect when a channel uses those options but has read the
entire object, there's relatively little benefit to be gained for
the added complexity.

I decided to attach the object checksums to the underlying
`ContentChannel` rather than the higher level
`GoogleCloudStorageGrpcReadChannel`. Opening a new `ContentChannel` is
relatively expensive, so if we're able to detect that a series of reads
across multiple `ContentChannel` instances amounts to a sequential read
of the whole object, we should avoid closing the channel in the first
place.

Change-Id: I176324cb4fea317a358eb0e56468cebf84c12063