Update API proto files for API readability changes
The API went through some substantial changes, and that's made it out
to prod now, so we need new proto files here!

While I was at it, I also figured out how to pull the default grpc target
from the service definition. The target is still defined in the code
(in the proto file), but at least this way we're not repeating
ourselves.

Change-Id: I39d5580048da071bd26b07d54da9582d1b974974