Clean up unused utilities and merge duplicated test utilities.
Make fields that don't need to be injected `final`.

Move mock utility into `testing` package.

Migrate throw error and throw exception response to inputstreamResponse
and use throwingInputStream.

For testing objects.Insert behavior, since it sends a header reqeust and
use the header in the response to build the insert request, intercepting
getContent is the best thing we can do since the returned content is not
read at all. This is cloudstorage api behavior so there is nothing we
can do about it.