Migrate tests from using Mockito mocks except for mockBatchHelper and related mocks (#737) (#743)
* Remove mickito mocks except for mockBatchHelper and related mocks.

Remove most of the mockito mocks. Mocks for Storage and requests are
completely removed. Only mocks for BatchHelper related objects remains.

Also moves the tests into their repective test class since they don't
use mockito now.

For testing objects.Insert behavior, since it sends a header reqeust and
use the header in the response to build the insert request, intercepting
getContent is the best thing we can do since the returned content is not
read at all. This is cloudstorage api behavior so there is nothing we
can do about it.

Co-authored-by: Igor Dvorzhak <idv@google.com>