chore: refactor GoogleCloudStorageClientReadChannel to introduce Footer class
Keeping track of the footer can be challening when only tracking
offsets. This change introduces a new inner class Footer, which
encapsulates the offsets, length and bytes related to the footer once it
has been read. Along with it, there are helper methods to make it easier
to perform operations related to the footer.